requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.14.0
annotated-types==0.7.0
anthropic==0.83.0
anyio==4.12.1
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.115.6
Flask==3.1.3
frozenlist==1.8.0
google-auth==2.48.0
google-genai==1.64.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.1
propcache==0.4.1
psycopg2-binary==2.9.10
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.0.1
python-jose[cryptography]==3.4.0
python-multipart==0.0.20
email-validator==2.2.0
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.32.1
websockets==15.0.1
Werkzeug==3.1.6
yarl==1.22.0
